requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langgraph>=0.0.40
openai>=1.10.0
requests>=2.28.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
googleapis-common-protos>=1.56.0
python-dateutil>=2.8.0

# AWS Dependencies
boto3>=1.34.0
botocore>=1.34.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0

# Scheduling
APScheduler>=3.10.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
tqdm>=4.66.0

# Frontend and Visualization
plotly>=5.17.0
streamlit-plotly-events>=0.0.6

# Text processing and fuzzy matching
# difflib is built-in to Python, no need to install
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
